pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.26.0",
    "litellm>=1.82.1",
    "python-dotenv>=1.2.2",
    "setuptools>=82.0.1",
]
[tool.setuptools.packages.find]
where = ["."]